// frontend/src/types/sale.ts
import { Seller } from "./seller"

export type Sale = {
    id: number
    vissited: number
    deals: number
    amount: number
    date: string
    seller: Seller
}

export type SalePage = {
    content?: Sale[]
    last: boolean
    totalElements: number
    totalPages: number
    size?: number
    number: number
    first: boolean
    numberOfElements?: number
    empty?: boolean
}

export type SaleSum = {
    sellerName: string
    sum: number
  }

export type  SaleSuccess = {
    sellerName: string
    visited: number
    deals: number
  }